[mod/#21] onBookItemClick 수정

+++ app/src/main/java/sopt/org/millie/presentation/search/model/SearchBookModel.kt
@@ -3,7 +3,7 @@ import androidx.compose.runtime.Immutable
 
 @Immutable
 data class SearchBookModel(
-    val bookId: Int,
+    val bookId: Long,
     val bookCoverImageUrl: String,
     val bookTitle: String,
     val bookAuthor: String,
